requirements_test.txt
# Testing dependencies for Serac
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
aiohttp>=3.9.0
